+++ Wikivoyage/LocationWebViewController.swift
@@ -10,62 +10,37 @@ import UIKit
 import WebKit
 import Alamofire
 import SwiftyJSON
+import MagicalRecord
 
-class LocationWebViewController: UIViewController, WKNavigationDelegate, UIGestureRecognizerDelegate {
+class LocationWebViewController: WebViewController {
 
-    var pageId: Int!
-    var pageTitle: String!
-    var webView: WKWebView!
-    @IBOutlet var progressView: UIProgressView!
-    
-    var style: String?
-    var zoom: String?
+    var originalURLSet: Bool = false
+    var originalURL: String?
+    @IBOutlet var saveButton: UIBarButtonItem!
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        setupWebView()
-        setupGestureRecognizer()
-        getScripts()
-        requestURL()
+        saveButton.enabled = false
     }
     
-    func setupWebView() {
-        webView = WKWebView()
-        webView.allowsBackForwardNavigationGestures = true
-        webView.navigationDelegate = self
+    override func webView(webView: WKWebView, didCommitNavigation navigation: WKNavigation!) {
+        super.webView(webView, didCommitNavigation: navigation)
         
-        self.view.addSubview(webView)
-        self.view.sendSubviewToBack(webView)
-        
-        // Edge contraints
-        webView.setTranslatesAutoresizingMaskIntoConstraints(false)
-        let height = NSLayoutConstraint(item: webView, attribute: .Height, relatedBy: .Equal, toItem: self.view, attribute: .Height, multiplier: 1, constant: 0)
-        let width = NSLayoutConstraint(item: webView, attribute: .Width, relatedBy: .Equal, toItem: self.view, attribute: .Width, multiplier: 1, constant: 0)
-        self.view.addConstraints([height, width])
-    }
-    
-    func setupGestureRecognizer() {
-        let gestureRecognizer = UITapGestureRecognizer(target: self, action: "webViewTapped:")
-        gestureRecognizer.delegate = self
-        gestureRecognizer.numberOfTapsRequired = 2
-        self.webView.addGestureRecognizer(gestureRecognizer)
-    }
-    
-    func getScripts() {
-        if let styleScriptURL = NSBundle.mainBundle().pathForResource("StyleScript", ofType: "js") {
-            style = String(contentsOfFile:styleScriptURL, encoding:NSUTF8StringEncoding, error: nil)
+        if !originalURLSet {
+            originalURL = webView.URL?.absoluteString
+            originalURLSet = true
         }
         
-        if let zoomScriptURL = NSBundle.mainBundle().pathForResource("ZoomScript", ofType: "js") {
-            zoom = String(contentsOfFile:zoomScriptURL, encoding:NSUTF8StringEncoding, error: nil)
+        if webView.URL?.absoluteString == originalURL {
+            saveButton.enabled = true
+        } else {
+            saveButton.enabled = false
         }
     }
     
-    func requestURL() {
-        let newTitle = pageTitle.stringByReplacingOccurrencesOfString(" ", withString: "_", options: nil, range: nil)
-        let url = NSURL(string: "http://en.m.wikivoyage.com/wiki/"+newTitle)
-        let request = NSURLRequest(URL: url!)
-        webView.loadRequest(request)
+    // Save page
+    @IBAction func savePage(sender: AnyObject) {
+        downloadText()
     }
     
     func downloadText() {
@@ -75,121 +50,24 @@ class LocationWebViewController: UIViewController, WKNavigationDelegate, UIGestu
                 NSLog("Error: \(error)")
             } else {
                 let json = JSON(data!)
-                let text = json["parse"]["text"]["*"]
+                
+                // Check if it's already saved and if currently displaying Wikivoyage page
+                let id = NSNumber(integer: self.pageId)
+                if let savedPage = SavedPage.MR_findFirstByAttribute("id", withValue: id) {
+                    savedPage.title = self.pageTitle
+                    savedPage.html = json["parse"]["text"]["*"].stringValue
+                } else {
+                    let savedPage = SavedPage.MR_createEntity()
+                    savedPage.title = self.pageTitle
+                    savedPage.id = self.pageId
+                    savedPage.html = json["parse"]["text"]["*"].stringValue
+                }
+                
+                NSManagedObjectContext.MR_defaultContext().MR_saveToPersistentStoreAndWait()
             }
         }
     }
     
-    // WebView delegate
-    func webView(webView: WKWebView, didStartProvisionalNavigation navigation: WKNavigation!) {
-        // Show nav bar when going to new page
-        self.navigationController?.setNavigationBarHidden(false, animated: true)
-    }
-    
-    func webView(webView: WKWebView, didCommitNavigation navigation: WKNavigation!) {
-        // Inject style and zoom CSS
-        if isHostWikiURL(webView.URL?.host) {
-            if style != nil { webView.evaluateJavaScript(style!, completionHandler: nil) }
-            if zoom != nil { webView.evaluateJavaScript(zoom!, completionHandler: nil) }
-        }
-    }
-    
-    func isHostWikiURL(url: String?) -> Bool {
-        if let components = url?.componentsSeparatedByString(".") {
-            if contains(components, "wikivoyage") || contains(components, "wikipedia") || contains(components, "wikimedia") && contains(components, "org") {
-                return true
-            } else {
-                return false
-            }
-        } else {
-            return false
-        }
-    }
-    
-    func webView(webView: WKWebView, didFailProvisionalNavigation navigation: WKNavigation!, withError error: NSError) {
-        // Error message
-        let alert = UIAlertController(title: "Error", message: error.localizedDescription, preferredStyle: .Alert)
-        alert.addAction(UIAlertAction(title: "OK", style: .Default, handler: nil))
-        presentViewController(alert, animated: true, completion: nil)
-    }
-    
-    func webView(webView: WKWebView, didFinishNavigation navigation: WKNavigation!) {
-        // Reset progress view after loading page
-        progressView.setProgress(0.0, animated: false)
-    }
-    
-    // Gesture recognizer
-    func gestureRecognizer(gestureRecognizer: UIGestureRecognizer, shouldRecognizeSimultaneouslyWithGestureRecognizer otherGestureRecognizer: UIGestureRecognizer) -> Bool {
-        return true
-    }
-    
-    func webViewTapped(recognizer: UITapGestureRecognizer) {
-        if let navigationController = self.navigationController {
-            let change = !navigationController.navigationBarHidden
-            navigationController.setNavigationBarHidden(change, animated: true)
-        }
-    }
-    
-    // Progress view and navigation controller
-    override func viewWillAppear(animated: Bool) {
-        super.viewWillAppear(animated)
-        
-        webView.addObserver(self, forKeyPath: "estimatedProgress", options: .New, context: nil)
-        webView.addObserver(self, forKeyPath: "title", options: .New, context: nil)
-        
-        self.navigationController?.hidesBarsOnSwipe = true
-    }
-    
-    // Disabling back swipe gesture only works in viewWillLayoutSubviews
-    override func viewWillLayoutSubviews() {
-        super.viewWillLayoutSubviews()
-        if self.navigationController != nil {
-            if self.navigationController!.respondsToSelector("interactivePopGestureRecognizer") {
-                self.navigationController?.interactivePopGestureRecognizer.enabled = false
-            }
-        }
-    }
-    
-    override func viewWillDisappear(animated: Bool) {
-        super.viewWillDisappear(animated)
-        
-        webView.removeObserver(self, forKeyPath: "estimatedProgress")
-        webView.removeObserver(self, forKeyPath: "title")
-        
-        // Reset navigation controller
-        self.navigationController?.setNavigationBarHidden(false, animated: true)
-        self.navigationController?.hidesBarsOnSwipe = false
-        if self.navigationController != nil {
-            if self.navigationController!.respondsToSelector("interactivePopGestureRecognizer") {
-                self.navigationController?.interactivePopGestureRecognizer.enabled = true
-            }
-        }
-    }
-    
-    override func observeValueForKeyPath(keyPath: String, ofObject object: AnyObject, change: [NSObject : AnyObject], context: UnsafeMutablePointer<Void>) {
-        if (keyPath == "estimatedProgress") {
-            // Bug where estimatedProgress = 0.1 even for pages that are already loaded
-            if webView.estimatedProgress > 0.1 && webView.estimatedProgress < 1.0 {
-                // Show progress if it's between 0.1 and 1.0
-                progressView.hidden = false
-                progressView.setProgress(Float(webView.estimatedProgress), animated: true)
-            } else {
-                progressView.hidden = true
-            }
-        }
-        
-        if (keyPath == "title") {
-            if let newTitle = webView.title?.stringByReplacingOccurrencesOfString(" – Travel guide at Wikivoyage", withString: "", options: nil, range: nil) {
-                self.title = newTitle
-            } else {
-                self.title = ""
-            }
-        }
-    }
-    
-    override func prefersStatusBarHidden() -> Bool {
-        return self.navigationController!.navigationBarHidden
-    }
     
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
