+++ Wikivoyage/OfflineScript.js
@@ -1,0 +1,4 @@
+var a = document.getElementsByTagName('a');
+for (var i = 0; i < a.length; i++) {
+    a[i].removeAttribute('href');
+}
