+++ Wikivoyage/SavedLocationWebViewController.swift
@@ -1,0 +1,11 @@
+//
+//  SavedLocationWebViewController.swift
+//  Wikivoyage
+//
+//  Created by Ben Meline on 9/16/15.
+//  Copyright (c) 2015 Ben Meline. All rights reserved.
+//
+
+class SavedLocationWebViewController: WebViewController {
+
+}
